+++ packages/create-mechanic/index.js
@@ -1,4 +1,5 @@
 const path = require("path");
+const fs = require("fs-extra");
 const inquirer = require("inquirer");
 const {
   spinners: { mechanicSpinner: spinner }
@@ -12,7 +13,9 @@ const {
   installationMethodQuestion,
   generateProjectTemplate,
   installDependencies,
-  checkLockFile
+  tryGitInit,
+  checkLockFile,
+  confirmGitQuestion
 } = require("./new-project");
 const {
   baseExists,
@@ -28,6 +31,20 @@ const sleep = (ms = 1000) => new Promise(resolve => setTimeout(resolve, ms));
 const nullishCoalescingOp = (arg1, arg2) => (arg1 != null ? arg1 : arg2);
 
 const askToInstall = async projectName => {
+  // Project directory
+  const cwd = path.resolve(projectName);
+
+  // List out dependencies being installed
+  const packageJsonPath = path.join(cwd, "package.json");
+  const packageObj = JSON.parse(await fs.readFile(packageJsonPath, "utf8"));
+  log(content.installingDependenciesMessage);
+  for (const depType of ["devDependencies", "dependencies"]) {
+    for (const dep in packageObj[depType]) {
+      log(content.dependencyItem(dep));
+    }
+  }
+  log();
+
   // Install dependencies in new project directory
   const { install } = await inquirer.prompt(confirmInstallQuestion);
   await sleep();
@@ -44,6 +61,18 @@ const askToInstall = async projectName => {
     return { success, installingMethod };
   }
   return install;
+};
+
+const askToInitGit = async projectName => {
+  // Install dependencies in new project directory
+  const { gitInit } = await inquirer.prompt(confirmGitQuestion);
+  await sleep();
+
+  if (gitInit) {
+    const success = await tryGitInit(projectName);
+    return { success };
+  }
+  return gitInit;
 };
 
 const command = async argv => {
@@ -106,6 +135,7 @@ const command = async argv => {
     const { confirmContinue } = await inquirer.prompt(confirmDFQuestion);
     await sleep();
     if (confirmContinue) {
+      log();
       log(content.designFunctionBasesDescription);
     } else {
       skipFunctions = true;
@@ -139,15 +169,26 @@ const command = async argv => {
       functionAnswers.functionName,
       functionQuestions[3].default
     );
-    await generateFunctionTemplate(projectName, {
-      typeOfBaseUsed: usesBase,
-      base: finalBase,
-      functionName
-    });
+    const config = require(path.join(
+      path.resolve(projectName),
+      "mechanic.config.js"
+    ));
+    await generateFunctionTemplate(
+      projectName,
+      {
+        typeOfBaseUsed: usesBase,
+        base: finalBase,
+        functionName
+      },
+      config
+    );
   }
 
   // Install dependencies in new project directory
   const install = await askToInstall(projectName);
+
+  // Try initializing git repository
+  await askToInitGit(projectName);
 
   // Done!
   log(content.doneAndNextStepsMessage(projectName, install));