+++ packages/create-mechanic/new-project.js
@@ -52,6 +52,15 @@ const installationMethodQuestion = [
   }
 ];
 
+const confirmGitQuestion = [
+  {
+    name: "gitInit",
+    type: "confirm",
+    message: content.confirmGitInitQuestion,
+    default: true
+  }
+];
+
 const generateProjectTemplate = async (projectName, typeOfBaseUsed) => {
   spinner.start(content.generateProjectStart);
 
@@ -63,7 +72,7 @@ const generateProjectTemplate = async (projectName, typeOfBaseUsed) => {
   // Copying content promises
   await Promise.all([
     // Copy array of files that get duplicated without change
-    ...["mechanic.config.js", "README.md"].map(filename =>
+    ...["mechanic.config.js", "README.md", "_gitignore"].map(filename =>
       fs.copyFile(
         path.join(projectTemplateDir, filename),
         path.join(directory, filename.replace(/^_/, "."))
@@ -88,6 +97,7 @@ const generateProjectTemplate = async (projectName, typeOfBaseUsed) => {
   ]);
 
   spinner.succeed(content.generateProjectSuccess(typeOfBaseUsed, projectName));
+  log();
   log(content.projectContents(path.dirname(directory)));
 };
 
@@ -126,12 +136,71 @@ const installDependencies = async (projectName, installingMethod) => {
   }
 };
 
+async function isInGitRepository(cwd) {
+  try {
+    await execa("git", ["rev-parse", "--is-inside-work-tree"], { cwd });
+    return true;
+  } catch (_) {}
+  return false;
+}
+
+async function isInMercurialRepository(cwd) {
+  try {
+    await execa("hg", ["--cwd", ".", "root"], { cwd });
+    return true;
+  } catch (_) {}
+  return false;
+}
+
+// Adaptation of https://github.com/vercel/next.js/blob/canary/packages/create-next-app/helpers/git.ts
+const tryGitInit = async projectName => {
+  // Project directory
+  const cwd = path.resolve(projectName);
+  let didInit = false;
+  try {
+    await execa("git", ["--version"], { cwd });
+    if (
+      (await isInGitRepository(cwd)) ||
+      (await isInMercurialRepository(cwd))
+    ) {
+      spinner.info(content.repositoryFound);
+      return false;
+    }
+
+    await execa("git", ["init"], { cwd });
+    didInit = true;
+
+    await execa("git", ["checkout", "-b", "main"], { cwd });
+
+    await execa("git", ["add", "-A"], { cwd });
+    await execa(
+      "git",
+      ["commit", "-m", "Initial commit from Create Mechanic"],
+      { cwd }
+    );
+    // Notify success
+    spinner.succeed(content.gitInitSucceed);
+    return true;
+  } catch (e) {
+    if (didInit) {
+      try {
+        fs.rmSync(path.join(cwd, ".git"));
+      } catch (_) {}
+    }
+    // Notify failure
+    spinner.fail(content.gitInitFailed);
+    return false;
+  }
+};
+
 module.exports = {
   getProjectQuestion,
   confirmDFQuestion,
   confirmInstallQuestion,
   installationMethodQuestion,
+  confirmGitQuestion,
   generateProjectTemplate,
   checkLockFile,
-  installDependencies
+  installDependencies,
+  tryGitInit
 };